refactor: separated constants into files by category, removed state from aboutme
