+++ lib/constants/navigationConstants.ts
@@ -1,0 +1,20 @@
+export const navLinks = [
+  { href: "#projects", label: "Projects" },
+  { href: "#blog", label: "Blog" },
+  { href: "#contact", label: "Contact" },
+];
+
+export const socialLinks = [
+  {
+    href: "https://www.linkedin.com/in/joshuatuddenham/",
+    icon: "/media/nav/linkedin-original.svg",
+    label: "LinkedIn",
+    id: "linkedin",
+  },
+  {
+    href: "https://github.com/joshuaisaact",
+    icon: "/media/nav/github.svg",
+    label: "GitHub",
+    id: "github",
+  },
+];
