+++ lib/constants/siteConfig.ts
@@ -1,0 +1,8 @@
+export const SITE_CONFIG = {
+  title: "Joshua Tuddenham",
+  description: "Full-stack engineer portfolio",
+  author: "Joshua Tuddenham",
+  url: "https://joshuatuddenham.com",
+  githubUrl: "https://github.com/joshuaisaact",
+  linkedinUrl: "https://www.linkedin.com/in/joshuatuddenham/",
+} as const;
