+++ lib/constants/themeConstants.ts
@@ -1,0 +1,10 @@
+export const THEME_COLORS = [
+  "#e07a14",
+  "#723e5a",
+  "#9c852a",
+  "#C589A3",
+  "#42d499",
+  "#5459de",
+  "#481682",
+  "#3e7256",
+] as const;
